% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_twitter.R
\name{analisis_twitter}
\alias{analisis_twitter}
\title{Análisis del Twitter en función de palabras clave.}
\usage{
analisis_twitter(palabras_clave)
}
\arguments{
\item{palabras_clave}{}
}
\value{
json
}
\description{
Analiza el Twitter en base a un vector de palabras clave.
}
\examples{
analisis_twitter("ermua,izarra centre,ayuntamiento de ermua")

}
