% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_twitter_ce.R
\name{analisis_twitter_ce}
\alias{analisis_twitter_ce}
\title{Análisis del Twitter cuentas.}
\usage{
analisis_twitter_ce(centros_empresa)
}
\arguments{
\item{centros_empresa}{}
}
\value{
json
}
\description{
Analiza el Twitter en base a un vector (datos) de cuentas/usuarios de Twitter especificados.
}
